@@ -68,7 +68,7 @@
     "rollup-plugin-replace": "^1.1.1",
     "rollup-watch": "^4.0.0",
     "selenium-server": "^2.53.1",
-    "typescript": "^2.0.3",
+    "typescript": "^2.5.3",
     "uglify-js": "^3.0.17",
     "vue": "^2.3.0",
     "vue-loader": "^12.0.2",
